+++ lilya/routing.py
@@ -361,7 +361,7 @@ class Path(BaseHandler, BasePath):
         if middleware is not None:
             self.middleware = [wrap_middleware(mid) for mid in middleware]
         else:
-            self.middleware = middleware if middleware is not None else []
+            self.middleware = middleware or []
 
         self.permissions = permissions if permissions is not None else []
         self.exception_handlers = {} if exception_handlers is None else dict(exception_handlers)
@@ -428,7 +428,9 @@ class Path(BaseHandler, BasePath):
                 self.exception_handlers.update(controller.exception_handlers)
 
             if controller.dependencies:
-                self.dependencies.update(controller.dependencies)
+                self.dependencies.update(
+                    {k: wrap_dependency(v) for k, v in controller.dependencies.items()}
+                )
 
             if controller.before_request:
                 self.before_request.extend(controller.before_request)
@@ -928,7 +930,10 @@ class Host(BasePath):
         self.host_regex, self.host_format, self.param_convertors, self.path_start = compile_path(
             host
         )
-        self.middleware = middleware if middleware is not None else []
+        if middleware is not None:
+            self.middleware = [wrap_middleware(mid) for mid in middleware]
+        else:
+            self.middleware = middleware or []
         self.permissions = permissions if permissions is not None else []
 
         # Wrap dependencies
@@ -1229,7 +1234,11 @@ class BaseRouter:
         self.include_in_schema = include_in_schema
         self.deprecated = deprecated
 
-        self.middleware = middleware if middleware is not None else []
+        if middleware is not None:
+            self.middleware = [wrap_middleware(mid) for mid in middleware]
+        else:
+            self.middleware = middleware or []
+
         self.permissions = permissions if permissions is not None else []
         self.settings_module = settings_module
         self.middleware_stack = self.app
@@ -2291,7 +2300,11 @@ class Include(BasePath):
 
         self.app = self.__base_app__
 
-        self.middleware = middleware if middleware is not None else []
+        if middleware is not None:
+            self.middleware = [wrap_middleware(mid) for mid in middleware]
+        else:
+            self.middleware = middleware or []
+
         self.permissions = permissions if permissions is not None else []
         self.exception_handlers = {} if exception_handlers is None else dict(exception_handlers)